Auth work is mostly complete.
If a username[:password] is provided with the --user flag, then that us used.

If no --user flag is provided, check for a token and use it if found.

If no token is found, then prompt for username (default to https user,
then system user with option to override) and password.

Just need to add a check for git credentials before the last step above
(if it is possible - see issue #1) and then auth work will be complete.